requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pillow==11.1.0
django-filter==25.1
PyJWT==2.9.0


